Make Cursor::setReadPreference() chainable and recreatable
Change the return value for setReadPreference() so it can be chained with other methods. The boolean return code from the MongoCursor method will be checked and used to throw an exception if necessary.

Save the current read preference and tags on Cursor, so it can be recreated (as is done for slaveOkay).