Wrap driver classes directly and remove Connection reinit logic
Constructor signatures have changed. Collections and Databases are now constructed with the MongoCollection and MongoDB instances to be wrapped, respectively. The getter methods for driver classes no longer operate like factories (returning new instances on each call), which makes it possible to set and maintain read preferences.

The ability to reconstruct MongoClient has been removed from Cursor::retry() and Connection::initialize(), since that is not necessary for the driver to reestablish a database connection.

For the Loggable classes, the loggerCallable argument has been moved to the end of the constructor. This removes previous inconsistencies between Loggable and non-Loggable constructors.