Don't throw InvalidArgumentException in Cursor::setReadPreference()
The driver may emit an E_WARNING for invalid arguments, but the return value should always be the MongoCursor (not a boolean).

Move the saving of readPreference and tags arguments after the driver method, so that we don't capture the arguments if an error handler decides to throw an exception.

See: https://jira.mongodb.org/browse/PHP-920