Keep EventArgs immutable and selectively use MutableEventArgs
This reverts earlier changes that added mutability to EventArgs (among other named event classes). In the interest of keeping EventArgs immutable, we'll now use MutableEventArgs for those methods that would benefit from it.

Note: preBatchInsert, preInsert, and preSave will continue to use immutable EventArgs for their inputs, due to the fact that the driver is prone to modifying the input arguments in memory (this is different than by-reference modification). If we allowed an event listener to mutate that argument, users would no longer be able to rely on "_id" injection on the value passed into the Collection method.